splitted the grouped alert types into multiple modules
