+++ elementary/monitor/alerts/grouped_alerts/__init__.py
@@ -1,0 +1,9 @@
+from .grouped_alert import GroupedAlert
+from .grouped_by_table import GroupedByTableAlerts
+from .grouping_type import GroupingType
+
+__all__ = [
+    "GroupedAlert",
+    "GroupedByTableAlerts",
+    "GroupingType",
+]
