+++ elementary/monitor/alerts/grouped_alerts/grouped_by_table.py
@@ -1,7 +1,6 @@
-from datetime import datetime
-from enum import Enum
 from typing import Dict, List, Optional, Union
 
+from elementary.monitor.alerts.grouped_alerts.grouped_alert import GroupedAlert
 from elementary.monitor.alerts.model_alert import ModelAlertModel
 from elementary.monitor.alerts.source_freshness_alert import SourceFreshnessAlertModel
 from elementary.monitor.alerts.test_alert import TestAlertModel
@@ -10,48 +9,6 @@ from elementary.monitor.data_monitoring.alerts.integrations.utils.report_link im
     get_model_test_runs_link,
 )
 from elementary.utils.models import get_shortened_model_name
-
-
-class GroupingType(Enum):
-    BY_ALERT = "alert"
-    BY_TABLE = "table"
-
-
-class GroupedAlert:
-    def __init__(
-        self,
-        alerts: List[Union[TestAlertModel, ModelAlertModel, SourceFreshnessAlertModel]],
-    ) -> None:
-        self.alerts = alerts
-        self.test_errors: List[Union[TestAlertModel, SourceFreshnessAlertModel]] = []
-        self.test_warnings: List[Union[TestAlertModel, SourceFreshnessAlertModel]] = []
-        self.test_failures: List[Union[TestAlertModel, SourceFreshnessAlertModel]] = []
-        self.model_errors: List[ModelAlertModel] = []
-        self._sort_alerts()
-
-    @property
-    def detected_at(self) -> datetime:
-        return min(alert.detected_at or datetime.max for alert in self.alerts)
-
-    @property
-    def status(self) -> str:
-        if self.model_errors or self.test_errors:
-            return "error"
-        elif self.test_failures:
-            return "failure"
-        else:
-            return "warn"
-
-    def _sort_alerts(self):
-        for alert in self.alerts:
-            if isinstance(alert, ModelAlertModel):
-                self.model_errors.append(alert)
-            elif alert.status == "error":
-                self.test_errors.append(alert)
-            elif alert.status == "warn":
-                self.test_warnings.append(alert)
-            else:
-                self.test_failures.append(alert)
 
 
 class GroupedByTableAlerts(GroupedAlert):
