+++ elementary/monitor/alerts/grouped_alerts/grouping_type.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class GroupingType(str, Enum):
+    BY_ALERT = "alert"
+    BY_TABLE = "table"
